[update] Reset Project
