+++ .prettierrc
@@ -1,0 +1,7 @@
+useTabs: false
+tabWidth: 2
+endOfLine: lf
+semi: true
+trailingComma: es5
+singleQuote: true
+printWidth: 120
