+++ build.gradle
@@ -2,15 +2,40 @@ plugins {
 	id 'java'
 	id 'org.springframework.boot' version '3.1.0'
 	id 'io.spring.dependency-management' version '1.1.0'
+	id 'com.diffplug.spotless' version '6.18.0'
 }
 
 group = 'kr.sanus'
-version = '0.0.2'
+version = '0.1.0'
 sourceCompatibility = '17'
 
 configurations {
 	compileOnly {
 		extendsFrom annotationProcessor
+	}
+}
+spotless{
+	format("yaml"){
+		target("**/*.yaml","**/*.yml")
+		prettier().configFile(".prettierrc")
+	}
+	java{
+		removeUnusedImports()
+		googleJavaFormat()
+		importOrder(
+				"java",
+				"jakarta",
+				"lombok",
+				"org.springframework",
+				"",
+				"\\#",
+				"kr.sanus.news",
+				"\\#kr.sanus.news"
+		)
+		indentWithTabs(2)
+		indentWithSpaces(2)
+		trimTrailingWhitespace()
+		endWithNewline()
 	}
 }
