+++ docker-compose.yml
@@ -1,4 +1,4 @@
-version: "3.3"
+version: '3.3'
 
 services:
   db_master:
@@ -14,4 +14,4 @@ services:
       - ./master_db/data:/var/lib/mysql
       - ./master_db/config/:/etc/mysql/conf.d
     ports:
-      - "3307:3306"
+      - '3307:3306'
