--- src/main/java/kr/sanus/news/base/HomeController.java
@@ -1,14 +1,0 @@
-package kr.sanus.news.base;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class HomeController {
-
-    @GetMapping("/")
-    public String home() {
-        return "index";
-    }
-
-}
