+++ src/main/java/kr/sanus/news/common/config/SecurityConfiguration.java
@@ -1,0 +1,50 @@
+package kr.sanus.news.common.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
+
+import static org.springframework.security.config.Customizer.withDefaults;
+
+@Configuration
+@EnableWebSecurity
+public class SecurityConfiguration {
+
+  @Bean
+  public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+    http.authorizeHttpRequests(
+        (authorize) ->
+          authorize
+            .requestMatchers("/admin/**")
+            .authenticated()
+            .requestMatchers("/member/detail")
+            .authenticated()
+            .anyRequest()
+            .permitAll())
+      .httpBasic(withDefaults())
+      .formLogin(
+        form ->
+          form.loginPage("/member/login")
+            .usernameParameter("email")
+            .passwordParameter("password")
+            .defaultSuccessUrl("/", true)
+            .permitAll())
+      .logout(
+        logout ->
+          logout
+            .logoutRequestMatcher(new AntPathRequestMatcher("/member/logout"))
+            .logoutSuccessUrl("/member/login")
+            .invalidateHttpSession(true));
+    return http.build();
+  }
+
+  @Bean
+  public PasswordEncoder passwordEncoder() {
+    return new BCryptPasswordEncoder();
+  }
+}
