+++ src/main/java/kr/sanus/news/common/controller/BaseController.java
@@ -1,0 +1,13 @@
+package kr.sanus.news.common.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class BaseController {
+
+  @GetMapping("/")
+  public String home() {
+    return "index";
+  }
+}
