+++ src/main/java/kr/sanus/news/common/entity/BaseEntity.java
@@ -1,11 +1,14 @@
-package kr.sanus.news.base;
+package kr.sanus.news.common.entity;
+
+import java.time.LocalDateTime;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
 import jakarta.persistence.MappedSuperclass;
-import java.time.LocalDateTime;
+
 import lombok.Getter;
 import lombok.Setter;
+
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
@@ -16,11 +19,11 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 @EntityListeners(AuditingEntityListener.class)
 public class BaseEntity {
 
-    public BaseEntity() {}
-    @CreatedDate
-    @Column(updatable = false)
-    private LocalDateTime createDate;
+  public BaseEntity() {}
 
-    @LastModifiedDate
-    private LocalDateTime updateDate;
+  @CreatedDate
+  @Column(updatable = false)
+  private LocalDateTime createDate;
+
+  @LastModifiedDate private LocalDateTime updateDate;
 }
