--- src/main/java/kr/sanus/news/member/MemberRole.java
@@ -1,15 +1,0 @@
-package kr.sanus.news.member;
-
-import lombok.Getter;
-
-@Getter
-public enum MemberRole {
-    ADMIN("ROLE_ADMIN"),
-    USER("ROLE_USER");
-
-    MemberRole(String value) {
-        this.value = value;
-    }
-
-    private final String value;
-}
