+++ src/main/java/kr/sanus/news/member/data/MemberRole.java
@@ -1,0 +1,15 @@
+package kr.sanus.news.member.data;
+
+import lombok.Getter;
+
+@Getter
+public enum MemberRole {
+  ADMIN("ROLE_ADMIN"),
+  USER("ROLE_USER");
+
+  MemberRole(String value) {
+    this.value = value;
+  }
+
+  private final String value;
+}
