+++ src/main/java/kr/sanus/news/member/data/entity/MemberEntity.java
@@ -1,4 +1,4 @@
-package kr.sanus.news.member;
+package kr.sanus.news.member.data.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -6,13 +6,16 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
-import kr.sanus.news.base.BaseEntity;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
+
+import kr.sanus.news.common.entity.BaseEntity;
+import kr.sanus.news.member.data.dto.MemberDto;
 
 @Entity
 @Getter
@@ -24,25 +27,20 @@ import lombok.ToString;
 @Table(name = "member")
 public class MemberEntity extends BaseEntity {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
 
-    @Column(unique = true)
-    private String email;
+  @Column(unique = true)
+  private String email;
 
-    private String password;
+  private String password;
 
-    private String name;
+  private String name;
 
-    private String mobile;
+  private String mobile;
 
-    public MemberDto toDto() {
-        return MemberDto.builder()
-            .id(id)
-            .email(email)
-            .name(name)
-            .mobile(mobile)
-            .build();
-    }
+  public MemberDto toDto() {
+    return MemberDto.builder().id(id).email(email).name(name).mobile(mobile).build();
+  }
 }
