+++ src/main/java/kr/sanus/news/member/data/repository/MemberRepository.java
@@ -1,9 +1,12 @@
-package kr.sanus.news.member;
+package kr.sanus.news.member.data.repository;
 
 import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import kr.sanus.news.member.data.entity.MemberEntity;
 
 public interface MemberRepository extends JpaRepository<MemberEntity, Long> {
 
-    Optional<MemberEntity> findByEmail(String email);
+  Optional<MemberEntity> findByEmail(String email);
 }
