+++ src/main/resources/templates/common/layout.html
@@ -15,7 +15,7 @@
 <nav th:replace="~{common/navbar :: navbarFragment}"></nav>
 <div th:block layout:fragment="content"></div>
 <script th:src="@{/js/bootstrap/bootstrap.min.js}"></script>
-<th:block layout:fragment="script"></th:block>
+<div th:block layout:fragment="script"></div>
 </body>
 
 </html>
