+++ src/main/resources/templates/member/detail.html
@@ -26,6 +26,9 @@
       <button type="button" class="btn btn-dark align-self-center"
               onclick="location.href='editForm.html'"
               th:onclick="|location.href='@{/member/edit}'|">수정</button>
+      <button type="button" class="btn btn-dark align-self-center"
+              onclick="location.href='passwordForm.html'"
+              th:onclick="|location.href='@{/member/pass}'|">비밀번호 변경</button>
     </div>
 </div>
